test(marketing): fold guarded template coverage into workflow suite
